Removed module/library installations from the notebooks, changed data_dir path, removed image validation in Q2 split_dataset_paths function
Having code to install kaggle and tensorflow within the notebooks was unnecessary so it has been removed.

The directory to the dataset has been changed as I have downloaded it locally.

The image validation within the split_dataset_paths function has been removed, as all of the files except two are images, and this removal speeds that code up by 2-3 orders of magnitude.